feat: Add server for the CI server
Motivation: The CI server needs logic for cloning the repo, checking out
the correct commit, and then running the tests. This is what this commit
will do.

Tests: Tested by running python3 ci_server/ci_logic.py to verify that
it's cloning, checking out the commit, logging, and removing the repo
when finished.

Resolves #11