feat: add GitHub API function, tests and documentation
Motivation:
The CI server needs to automate how it creates and updates commit statuses for the commits that are evaluated.

This code connects to GitHubs API function "create commit status" which handles this automatically. Tests are also added to evaluate that all different functionalities of this API endpoint works. Documentation is also added to explain how to set this up on the CI server and locally for GitHub token authentication.

Resolves #3