feat: add GitHub API function, tests and documentation (#20)
* feat: add GitHub API function, tests and documentation

Motivation:
The CI server needs to automate how it creates and updates commit statuses for the commits that are evaluated.

This code connects to GitHubs API function "create commit status" which handles this automatically. Tests are also added to evaluate that all different functionalities of this API endpoint works. Documentation is also added to explain how to set this up on the CI server and locally for GitHub token authentication.

test: Move API function tests to test_ci_server folder

Move the test_GitHubAPI_functions.py file away from the ci_server folder and into the test_server folder in order to have all tests in the same folder

Resolves #3


---------

Co-authored-by: Ludvig Skare <lskare@kth.se>